Remove unused sync version of export/upload blobs
