+++ pdsmigration-web/src/api/jobs.rs
@@ -1,82 +1,16 @@
-use crate::api::{ExportBlobsApiRequest, UploadBlobsApiRequest};
 use crate::background_jobs::{JobManager, JobRecord};
-use crate::config::AppConfig;
-use crate::errors::{ApiError, ApiErrorBody};
-use crate::{post, Json};
+use crate::errors::ApiError;
+use crate::post;
 use actix_web::{get, web, HttpResponse};
-use pdsmigration_common::{ExportBlobsRequest, UploadBlobsRequest};
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 use uuid::Uuid;
 
+/// Response returned by all endpoints that enqueue a background job.
 #[derive(Debug, Deserialize, Serialize, ToSchema)]
 pub struct EnqueueJobResponse {
     #[schema(example = "550e8400-e29b-41d4-a716-446655440000")]
     pub job_id: String,
-}
-
-#[utoipa::path(
-    post,
-    path = "/jobs/export-blobs",
-    request_body = ExportBlobsApiRequest,
-    responses(
-        (status = 202, description = "Job enqueued", body = EnqueueJobResponse, content_type = "application/json"),
-        (status = 400, description = "Invalid request", body = ApiErrorBody, content_type = "application/json"),
-        (status = 401, description = "Authentication error", body = ApiErrorBody, content_type = "application/json"),
-        (status = 429, description = "Rate limit exceeded", body = ApiErrorBody, content_type = "application/json"),
-    ),
-    tag = "pdsmigration-web"
-)]
-#[tracing::instrument(skip(jobs, req))]
-#[post("/jobs/export-blobs")]
-pub async fn enqueue_export_blobs_job_api(
-    jobs: web::Data<JobManager>,
-    req: Json<ExportBlobsApiRequest>,
-) -> Result<HttpResponse, ApiError> {
-    let req_inner = req.into_inner();
-    let did = req_inner.did.clone();
-    tracing::info!("[{}] Enqueueing export-blobs job", did);
-    let id = jobs
-        .spawn_export_blobs(ExportBlobsRequest::from(req_inner))
-        .await?;
-    tracing::info!("[{}] Enqueued export-blobs job {}", did, id);
-    Ok(HttpResponse::Accepted().json(EnqueueJobResponse {
-        job_id: id.to_string(),
-    }))
-}
-
-#[utoipa::path(
-    post,
-    path = "/jobs/upload-blobs",
-    request_body = UploadBlobsApiRequest,
-    responses(
-        (status = 202, description = "Job enqueued", body = EnqueueJobResponse, content_type = "application/json"),
-        (status = 400, description = "Invalid request", body = ApiErrorBody, content_type = "application/json"),
-        (status = 401, description = "Authentication error", body = ApiErrorBody, content_type = "application/json"),
-        (status = 429, description = "Rate limit exceeded", body = ApiErrorBody, content_type = "application/json"),
-    ),
-    tag = "pdsmigration-web"
-)]
-#[tracing::instrument(skip(jobs, req, config))]
-#[post("/jobs/upload-blobs")]
-pub async fn enqueue_upload_blobs_job_api(
-    jobs: web::Data<JobManager>,
-    config: web::Data<AppConfig>,
-    req: Json<UploadBlobsApiRequest>,
-) -> Result<HttpResponse, ApiError> {
-    let req_inner = req.into_inner();
-    let did = req_inner.did.clone();
-    tracing::info!("[{}] Enqueueing upload-blobs job", did);
-    let id = jobs
-        .spawn_upload_blobs(
-            UploadBlobsRequest::from(req_inner),
-            config.server.concurrent_tasks_per_job,
-        )
-        .await?;
-    tracing::info!("[{}] Enqueued upload-blobs job {}", did, id);
-    Ok(HttpResponse::Accepted().json(EnqueueJobResponse {
-        job_id: id.to_string(),
-    }))
 }
 
 #[utoipa::path(
