+++ pdsmigration-web/src/main.rs
@@ -7,10 +7,9 @@ mod openapi;
 
 use crate::api::{
     activate_account_api, cancel_job_api, create_account_api, deactivate_account_api,
-    enqueue_export_blobs_job_api, enqueue_upload_blobs_job_api, export_blobs_api, export_pds_api,
-    get_job_api, get_service_auth_api, health_check, import_pds_api, list_jobs_api,
-    migrate_plc_api, migrate_preferences_api, missing_blobs_api, request_token_api,
-    upload_blobs_api,
+    enqueue_export_blobs_job_api, enqueue_upload_blobs_job_api, export_pds_api, get_job_api,
+    get_service_auth_api, health_check, import_pds_api, list_jobs_api, migrate_plc_api,
+    migrate_preferences_api, missing_blobs_api, request_token_api,
 };
 use crate::background_jobs::JobManager;
 use crate::config::AppConfig;
@@ -67,8 +66,6 @@ fn init_http_server(app_config: AppConfig) -> io::Result<Server> {
             .service(export_pds_api)
             .service(import_pds_api)
             .service(missing_blobs_api)
-            .service(export_blobs_api)
-            .service(upload_blobs_api)
             .service(enqueue_export_blobs_job_api)
             .service(enqueue_upload_blobs_job_api)
             .service(list_jobs_api)
@@ -175,8 +172,6 @@ mod tests {
                 .service(export_pds_api)
                 .service(import_pds_api)
                 .service(missing_blobs_api)
-                .service(export_blobs_api)
-                .service(upload_blobs_api)
                 .service(activate_account_api)
                 .service(deactivate_account_api)
                 .service(migrate_preferences_api)
