+++ pdsmigration-web/src/openapi.rs
@@ -11,12 +11,10 @@ use crate::api::*;
         activate_account_api,
         create_account_api,
         deactivate_account_api,
-        export_blobs_api,
         export_pds_api,
         import_pds_api,
         missing_blobs_api,
         request_token_api,
-        upload_blobs_api,
         migrate_preferences_api,
         migrate_plc_api,
         get_service_auth_api,
